split the indiserver.cpp into multiple files
