+++ indiserver/CMakeLists.txt
@@ -7,7 +7,21 @@ else()
     find_package(Threads REQUIRED)
     find_package(Libev REQUIRED)
 
-    add_executable(${PROJECT_NAME} indiserver.cpp)
+    add_executable(${PROJECT_NAME} indiserver.cpp
+                                   LocalDvrInfo.cpp
+                                   RemoteDvrInfo.cpp
+                                   UnixServer.cpp
+                                   TcpServer.cpp
+                                   Fifo.cpp
+                                   ClInfo.cpp
+                                   DvrInfo.cpp
+                                   MsgQueue.cpp
+                                   SerializedMsg.cpp
+                                   SerializedMsgWithoutSharedBuffer.cpp
+                                   SerializedMsgWithSharedBuffer.cpp
+                                   MsgChunck.cpp
+                                   Msg.cpp
+                                   Utils.cpp)
 
     target_link_libraries(indiserver indicore ${CMAKE_THREAD_LIBS_INIT} ${LIBEV_LIBRARIES})
     target_include_directories(indiserver SYSTEM PRIVATE ${LIBEV_INCLUDE_DIRS})
